feat: share auth_provider setting and add Entra group sync config
